refactor: Application에서 객체 생성 책임 분리 및 Controller 도입
