+++ src/test/java/application/SetUpParseTest.java
@@ -11,7 +11,7 @@ public class SetUpParseTest {
     @Test
     @DisplayName("상차림 번호 1을 왼상차림으로 변환한다.")
     void findOne() {
-        SetUp setUp = Application.toSetUp(1);
+        SetUp setUp = GameController.toSetUp(1);
 
         assertThat(setUp).isEqualTo(SetUp.LEFT_ELEPHANT);
     }
@@ -19,7 +19,7 @@ public class SetUpParseTest {
     @Test
     @DisplayName("상차림 번호 2를 오른상차림으로 변환한다.")
     void findTwo() {
-        SetUp setUp = Application.toSetUp(2);
+        SetUp setUp = GameController.toSetUp(2);
 
         assertThat(setUp).isEqualTo(SetUp.RIGHT_ELEPHANT);
     }
@@ -27,7 +27,7 @@ public class SetUpParseTest {
     @Test
     @DisplayName("상차림 번호 3을 안상차림으로 변환한다.")
     void findThree() {
-        SetUp setUp = Application.toSetUp(3);
+        SetUp setUp = GameController.toSetUp(3);
 
         assertThat(setUp).isEqualTo(SetUp.INNER_ELEPHANT);
     }
@@ -35,7 +35,7 @@ public class SetUpParseTest {
     @Test
     @DisplayName("상차림 번호 4를 바깥상차림으로 변환한다.")
     void findFour() {
-        SetUp setUp = Application.toSetUp(4);
+        SetUp setUp = GameController.toSetUp(4);
 
         assertThat(setUp).isEqualTo(SetUp.OUTER_ELEPHANT);
     }
@@ -43,7 +43,7 @@ public class SetUpParseTest {
     @Test
     @DisplayName("상차림 번호가 1에서 4 사이가 아니면 예외가 발생한다.")
     void invalidSetUpNumber() {
-        assertThatThrownBy(() -> Application.toSetUp(5))
+        assertThatThrownBy(() -> GameController.toSetUp(5))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 상차림 번호는 1, 2, 3, 4 중 하나여야 합니다.");
     }
